pyproject.toml
[tool.poetry]
name = "I2VGKCDM"
version = "0.1.0"
description = ""
authors = ["Ulugbek Shernazarov"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.15.2"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
diffusers = ">=0.19.0"
transformers = ">=4.30.0"
pillow = ">=10.0.0"
tqdm = ">=4.65.0"
opencv-python = ">=4.7.0"
streamlit = ">=1.25.0"
numpy = ">=1.24.0"
mlflow = ">=2.6.0"
comet-ml = ">=3.33.0"
python-dotenv = ">=1.0.0"
scikit-learn = ">=1.3.0"
matplotlib = ">=3.7.0"
tensorboard = ">=2.13.0"
wandb = ">=0.15.0"
librosa = "^0.11.0"
keyframed = "^0.3.15"
kornia = "^0.8.0"
compel = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
